Video page/content type

+++ js/imageman/dashboard.trash.js
@@ -1,11 +1,11 @@
 var $ = require('jquery');
 
 $(function() {
-	/**
-	 * 
-	 * CACHE DOM
-	 * 
-	 */
+/**
+ * 
+ * CACHE DOM
+ * 
+ */
 	var $contentList = $('#contentList'),
 	$contentTypeFilter = $contentList.find('select#filterContentList'),
 	$contentTbody = $contentList.find('tbody'),
@@ -15,134 +15,56 @@ $(function() {
 	$checkAll = $trashList.find('#trashCheckAll'),
 	$mainNav = $('#mainNav').children('ul.navbar-nav');
 
-	/**
-	 * 
-	 * BIND EVENTS
-	 * 
-	 */
+/**
+ * 
+ * BIND EVENTS
+ * 
+ */
 	// Trash button clicked
-	$contentList.on('click', '.trashContent', function(ev) {
-		ev.preventDefault();
+	$contentList.on('click', '.trashContent', trashContent);
+
+	// Delete button
+	$trashList.on('click', '.deleteContent', deleteContent);
+
+	// Restore button
+	$trashList.on('click', '.restoreContent', restoreContent);
+
+	// Empty Trash
+	$trashList.on('click', '.emptyTrash', emptyTrash);
+
+	// Delete Selected
+	$trashList.on('click', '.deleteSelected', deleteSelected);
+
+	// Restore Selected
+	$trashList.on('click', '.restoreSelected', restoreSelected);
+
+	// Check/uncheck-all behavior
+	$checkAll.on('change', checkAll);
+	$trashList.on('change', '.trashCheck', updateCheckboxes);
+
+	// Trash reload event
+	events.on('reloadTrash', reloadTrash);
+
+/**
+ * 
+ * MAIN FUNCTIONS
+ * 
+ */
+ 	
+ 	// Trash content
+ 	function trashContent(ev) {
+ 		ev.preventDefault();
 		var contentID = $(this).attr('id'),
 		$thisRow = $(this).closest('tr'),
 		confirmMessage;
-
 		if($thisRow.hasClass('page')) {
 			confirmMessage = 'Are you sure you want to trash this page? Associated content and subpages will be orphaned';
 		} else {
 			confirmMessage = 'Are you sure you want to trash this item?';
 		}
-		if(confirm(confirmMessage)) {
-			trashContent(contentID);
-		}	
-	});
-
-	// Delete button
-	$trashList.on('click', '.deleteContent', function(ev) {
-		ev.preventDefault();
-		var contentID = $(this).attr('id'),
-		$thisRow = $(this).closest('tr');
-
-		if($thisRow.hasClass('page')) {
-			confirmMessage = 'Are you sure you want to PERMANENTLY DELETE this page? Associated content and subpages will also be deleted. This action cannot be undone.';
-		} else {
-			confirmMessage = 'Are you sure you want to PERMANENTLY DELETE this item?';
+		if(!confirm(confirmMessage)) {
+			return false;
 		}
-		if(confirm(confirmMessage)) {
-			deleteContent(contentID, $thisRow);
-		}	
-	});
-
-	// Restore button
-	$trashList.on('click', '.restoreContent', function(ev) {
-		ev.preventDefault();
-		var contentID = $(this).attr('id'),
-		$thisRow = $(this).closest('tr');
-
-		restoreContent(contentID, $thisRow);
-	});
-
-	// Empty Trash
-	$trashList.on('click', '.emptyTrash', function(ev) {
-		ev.preventDefault();
-		if(confirm('Are you sure you want to PERMANENTLY DELETE all items in trash?')) {
-			emptyTrash();
-		}
-	});
-
-	// Delete Selected
-	$trashList.on('click', '.deleteSelected', function(ev) {
-		ev.preventDefault();
-		var checkedItems = buildCheckedArray();
-		if(checkedItems.length > 0) {
-			if(confirm('Are you sure you want to PERMANENTLY DELETE the selected items?')) {
-				deleteSelected(checkedItems);
-			}	
-		}
-	});
-
-	// Restore Selected
-	$trashList.on('click', '.restoreSelected', function(ev) {
-		ev.preventDefault();
-		var checkedItems = buildCheckedArray();
-		if(checkedItems.length > 0) {
-			if(confirm('Are you sure you want to restore the selected items?')) {
-				restoreSelected(checkedItems);
-			}	
-		}
-	});
-
-	// Check/uncheck-all behavior
-	$checkAll.on('change', function() {
-		var $trashChecks = $trashList.find('.trashCheck');
-		$trashChecks.each(function(){
-			$(this).prop('checked', $checkAll.prop('checked'));
-		});
-	});
-
-	$trashList.on('change', '.trashCheck', function() {
-		var $trashChecks = $trashList.find('.trashCheck');
-		// If user unchecks a box and the check-all box is checked, uncheck it!
-		if(!$(this).prop('checked') && $checkAll.prop('checked')) {
-			$checkAll.prop('checked', false);
-		}
-		// If all checkboxes are checked and check-all box is unchecked, check it!
-		if($('.trashCheck:checked').length == $trashChecks.length && !$checkAll.prop('checked')) {
-			$checkAll.prop('checked', true);
-		}
-	});
-
-	// Trash reload event
-	events.on('reloadTrash', function() {
-		reloadTrash();
-	});
-
-	/**
-	 * 
-	 * MAIN FUNCTIONS
-	 * 
-	 */
-   // Empty Trash
-   function emptyTrash()
-   {
-   	$.ajax({
-   		type: 'DELETE',
-   		url: baseURL + 'dashboard/emptyTrash/',
-   		dataType: 'json',
-   		success: function(data) {
-   			if(!data.error) {
- 					$trashTbody.find('tr').fadeOut(300, function() {
-			 			$(this).remove();
-			 		});
-			 		reloadTrash();
- 				}
-   		}
-   	});
-   }
-
- 	// Trash content
- 	function trashContent(contentID)
- 	{
  		$.ajax({
  			type: 'DELETE',
  			url: baseURL + 'dashboard/trashContent/' + contentID,
@@ -163,9 +85,41 @@ $(function() {
  		});
  	}
 
+   // Empty Trash
+   function emptyTrash(ev) {
+   	ev.preventDefault();
+		if(!confirm('Are you sure you want to PERMANENTLY DELETE all items in trash?')) {
+			return false;
+		}
+   	$.ajax({
+   		type: 'DELETE',
+   		url: baseURL + 'dashboard/emptyTrash/',
+   		dataType: 'json',
+   		success: function(data) {
+   			if(!data.error) {
+ 					$trashTbody.find('tr').fadeOut(300, function() {
+			 			$(this).remove();
+			 		});
+			 		reloadTrash();
+ 				}
+   		}
+   	});
+   }
+
  	// Delete content
- 	function deleteContent(contentID, $thisRow)
- 	{
+ 	function deleteContent(ev) {
+ 		ev.preventDefault();
+		var contentID = $(this).attr('id'),
+		$thisRow = $(this).closest('tr');
+
+		if($thisRow.hasClass('page')) {
+			confirmMessage = 'Are you sure you want to PERMANENTLY DELETE this page? Associated content and subpages will also be deleted. This action cannot be undone.';
+		} else {
+			confirmMessage = 'Are you sure you want to PERMANENTLY DELETE this item?';
+		}
+		if(!confirm(confirmMessage)) {
+			return false;
+		}
  		$.ajax({
  			type: 'DELETE',
  			url: baseURL + 'dashboard/deleteContent/' + contentID,
@@ -184,8 +138,16 @@ $(function() {
  	}
 
  	// Delete Selected
- 	function deleteSelected(checkedItems)
- 	{
+ 	function deleteSelected(ev) {
+ 		ev.preventDefault();
+		var checkedItems = buildCheckedArray();
+		if(checkedItems.length === 0) {
+			return false;
+		}
+		if(!confirm('Are you sure you want to PERMANENTLY DELETE the selected items?')) {
+			return false;
+		}	
+
  		$.ajax({
  			type: 'POST',
  			url: baseURL + 'dashboard/deleteMultiple/',
@@ -207,8 +169,10 @@ $(function() {
  	}
 
  	// Restore content
- 	function restoreContent(contentID, $thisRow)
- 	{
+ 	function restoreContent(ev) {
+ 		ev.preventDefault();
+		var contentID = $(this).attr('id'),
+		$thisRow = $(this).closest('tr');
  		$.ajax({
  			type: 'POST',
  			url: baseURL + 'dashboard/restoreContent/' + contentID,
@@ -229,8 +193,15 @@ $(function() {
  	}
 
  	// Restore Selected
- 	function restoreSelected(checkedItems)
- 	{
+ 	function restoreSelected(ev) {
+ 		ev.preventDefault();
+		var checkedItems = buildCheckedArray();
+		if(checkedItems.length === 0) {
+			return false;
+		}
+		if(!confirm('Are you sure you want to restore the selected items?')) {
+			return false;
+		}	
  		$.ajax({
  			type: 'POST',
  			url: baseURL + 'dashboard/restoreMultiple/',
@@ -252,6 +223,26 @@ $(function() {
  			}
  		});
  	}
+
+	// Check/uncheck All
+	function checkAll() {
+		var $trashChecks = $trashList.find('.trashCheck');
+		$trashChecks.each(function(){
+			$(this).prop('checked', $checkAll.prop('checked'));
+		});
+	}
+	// Update checkboxes
+	function updateCheckboxes() {
+		var $trashChecks = $trashList.find('.trashCheck');
+		// If user unchecks a box and the check-all box is checked, uncheck it!
+		if(!$(this).prop('checked') && $checkAll.prop('checked')) {
+			$checkAll.prop('checked', false);
+		}
+		// If all checkboxes are checked and check-all box is unchecked, check it!
+		if($('.trashCheck:checked').length == $trashChecks.length && !$checkAll.prop('checked')) {
+			$checkAll.prop('checked', true);
+		}
+	}
 
 	/**
 	 *
